Move ReactionRole mapping logic to DTO
