+++ src/UrfRidersBot.Persistence/DTOs/ReactionRoleDTO.cs
@@ -1,4 +1,9 @@
-﻿namespace UrfRidersBot.Persistence.DTOs
+﻿using DSharpPlus;
+using DSharpPlus.Entities;
+using UrfRidersBot.Core.Entities;
+using UrfRidersBot.Persistence.Mappers;
+
+namespace UrfRidersBot.Persistence.DTOs
 {
     public class ReactionRoleDTO
     {
@@ -11,6 +16,20 @@
             MessageId = messageId;
             Emoji = emoji;
             RoleId = roleId;
+        }
+        public ReactionRole ToDiscord(DiscordClient client, DiscordMessage message)
+        {
+            var guild = message.Channel.Guild;
+            var role = guild.GetRole(RoleId);
+            var emoji = EmojiMapper.ToDiscord(client, Emoji);
+
+            return new ReactionRole(message, emoji, role);
+        }
+
+        public static ReactionRoleDTO FromDiscord(ReactionRole reactionRole)
+        {
+            var emoji = EmojiMapper.FromDiscord(reactionRole.Emoji);
+            return new ReactionRoleDTO(reactionRole.Message.Id, emoji, reactionRole.Role.Id);
         }
     }
 }
